compiler: implement runtime-allocated arrays
This implements allocation of array with fixed size at
runtime. Register zero now points to the start of the "heap", which
was naivelly put on the middle of the address space for simplicity. On
the future, a pre-pass could be added to the compiler to find maximum
size of each static scope and optimize where the dynamic allocations
occur.

Also, those arrays don't ever get deallocated when exitting a scope.