replace walk interpreter with bytecode VM
This solves the effort duplication issue of having to implement things twice,
once for the walking intepreter and then again for the bytecode compiler. By
compiling and the interpreting the compiled bytecode, only the compiler needs to
know how to make sense of the AST. The walking interpreter is still available,
for now, through the `-w' flag.