rework strings and write builtin
Strings are no longer a type of operand and are instead pointers to a static
buffer in memory, terminated by setinel null character (like in C). The result
of compiling a string literal is a pointer to a statically allocated buffer
containing the characters of the string at the end of the dynamic allocation section.